Fixed conversion utils to convert Circle to Pirple world files
Fixed a bug with shop conversion of items that shopkeepers will
purchase. Changed conversion utilities to convert vnums and znums down
to an ordered set of zones and to convert vnums in item/shop/room/mob
files to pairs of values for zone number and internal virtual number for
each zone. IE: Item number 12000 is now zone under zone 120, item vnum
0. Removed top and bottom variables from zone. Zones now track the
number of each type of object they have, and there is no global version
of a room list etc. Zones should now be easily able to have an unlimited
number of each instead of a hard set 0-99 for vnums of each type.